Refactored Project by using an ImageHelper class
