+++ Medusa/ImageGenerator.py
@@ -14,6 +14,9 @@ def create_n_images(n):
         batch.append(create_image())
     return numpy.asarray(batch)
 
+def __propFilter(value,prop):
+    return value if(numpy.random.rand()<prop) else 0
+
 def create_image_with_color_prop(width = 64, height = 64, prop_red = 33, prop_green = 34, prop_blue = 33):
   return create_img_from_bytearray(create_bytearray_with_color_prop(width, height, prop_red, prop_green, prop_blue), 'RGBA')
 
@@ -35,18 +38,6 @@ def _create_blue_image_part(prop):
 def create_bytearray_with_color_prop(width = 64, height = 64, prop_red = 1, prop_green = 1, prop_blue = 1):
     return _create_red_image_part(prop_red)+_create_blue_image_part(prop_blue)+_create_green_image_part(prop_green)
 
-def __propFilter(value,prop):
-    return value if(numpy.random.rand()<prop) else 0
-
-    
-def change_brigthness_of_img(image, brightnessFactor=1.0):
-    changedImage = ImageEnhance.Brightness(image).enhance(brightnessFactor)
-    return changedImage
-
-def change_contrast_of_img(image, contrastFactor=1.0):
-    changedImage = ImageEnhance.Contrast(image).enhance(contrastFactor)
-    return changedImage
-
 def create_img_from_bytearray(bytearr, colorscheme='RGBA'):
     im=Image.fromarray(bytearr).convert(colorscheme)
     return im
@@ -57,8 +48,4 @@ def create_bytearray(width=64,height=64):
     casted = imarray.astype('uint8')
     return casted
 
-def img_to_bytearray(Image):
-    imgByteArr = io.BytesIO()
-    Image.save(imgByteArr, format='PNG')
-    imgByteArr = imgByteArr.getvalue()
-    return imgByteArr
+
